Remove new Person from RSA to Homescreen
