+++ GUI/src/app/homepage/startseite.component.html
@@ -15,3 +15,8 @@
         </mat-expansion-panel>
     </div>
 </mat-accordion>
+<div class="element-below-accordion-container">
+    <button mat-fab color="primary" (click)="openNameInputDialog()">
+        <mat-icon>person_add</mat-icon>
+    </button>
+</div>
