+++ GUI/src/app/rsa/rsa.component.html
@@ -71,8 +71,3 @@
         </mat-action-row>
     </mat-expansion-panel>
 </mat-accordion>
-<div class="element-below-accordion-container">
-    <button mat-fab color="primary" (click)="openNameInputDialog()">
-        <mat-icon>person_add</mat-icon>
-    </button>
-</div>
